requirements.txt
amqp==2.3.2
apipkg==1.5
asn1crypto==0.24.0
atomicwrites==1.1.5
attrs==18.1.0
Babel==2.6.0
bcrypt==3.1.4
billiard==3.5.0.4
celery==4.2.1
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
coverage==4.5.1
cryptography==2.3.1
Django==2.0
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-cors-headers==2.4.0
django-redis==4.9.0
djangorestframework==3.8.2
execnet==1.5.0
Fabric3==1.14.post1
flower==0.9.2
idna==2.7
JPype1==0.6.3
kombu==4.2.1
konlpy==0.5.1
more-itertools==4.2.0
nltk==3.3
paramiko==2.4.1
pluggy==0.6.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
py==1.5.4
pyasn1==0.4.4
pycparser==2.18
PyNaCl==1.2.1
pytest==3.6.3
pytest-cov==2.5.1
pytest-django==3.3.2
pytest-forked==0.2
pytest-xdist==1.22.3
pytz==2018.5
raven==6.9.0
redis==2.10.6
requests==2.19.1
six==1.11.0
tornado==5.1
urllib3==1.23
vine==1.1.4
